requirements.txt
absl-py==0.11.0
amqp==5.0.5
asgiref==3.3.0
astunparse==1.6.3
bert==2.2.0
bert-for-tf2==0.14.7
bert-tensorflow==1.0.4
billiard==3.6.3.0
cachetools==4.1.1
celery==5.0.5
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.2.1
dash-html-components==1.1.2
Django==3.1.3
django-crispy-forms==1.9.2
django-webpush==0.3.3
djangorestframework==3.12.2
erlastic==2.0.0
Flask==1.1.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
http-ece==1.1.0
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Preprocessing==1.1.2
kombu==5.0.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.2
params-flow==0.8.2
plotly==4.14.1
prompt-toolkit==3.0.16
protobuf==3.13.0
py-params==0.9.7
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
pywebpush==1.9.4
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.51.0
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0